fixed_sponsors manually re-applied commit 409c96491b2b7423338018b2acf5bec876e6e9d3 as there was something broken there
