+++ src/main/java/co/inajar/oursponsors/controllers/fec/FECController.java
@@ -2,7 +2,7 @@ package co.inajar.oursponsors.controllers.fec;
 
 import co.inajar.oursponsors.models.fec.SponsorRequest;
 import co.inajar.oursponsors.models.fec.SponsorResponse;
-import co.inajar.oursponsors.services.fec.SponsorsManager;
+import co.inajar.oursponsors.services.fec.SponsorManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,14 +21,14 @@ import java.util.stream.Collectors;
 public class FECController {
 
     @Autowired
-    private SponsorsManager sponsorsManager;
+    private SponsorManager sponsorManager;
 
     @PreAuthorize("isAuthenticated()")
     @PostMapping(path = "get_sponsors")
     public ResponseEntity<List<SponsorResponse>> getSponsors(@RequestBody SponsorRequest data) {
         var response = new ArrayList<SponsorResponse>();
         var httpStatus = HttpStatus.OK;
-        var possibleSponsors = sponsorsManager.getSponsors(data);
+        var possibleSponsors = sponsorManager.getSponsors(data);
         if (!possibleSponsors.isEmpty()) {
             var list = possibleSponsors.parallelStream()
                     .map(SponsorResponse::new)
