+++ src/main/java/co/inajar/oursponsors/controllers/opensecrets/AdminOpensecretsController.java
@@ -4,7 +4,7 @@ import co.inajar.oursponsors.models.opensecrets.CampaignResponse;
 import co.inajar.oursponsors.models.opensecrets.CommitteeRequest;
 import co.inajar.oursponsors.models.opensecrets.contributor.ContributorResponse;
 import co.inajar.oursponsors.models.opensecrets.sector.SectorResponse;
-import co.inajar.oursponsors.services.opensecrets.CandidatesApiManager;
+import co.inajar.oursponsors.services.opensecrets.CandidateApiManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,7 +18,7 @@ import java.util.List;
 public class AdminOpensecretsController {
 
     @Autowired
-    private CandidatesApiManager candidatesApiManager;
+    private CandidateApiManager candidateApiManager;
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping(path = "download_sectors/{part}")
@@ -27,8 +27,8 @@ public class AdminOpensecretsController {
         // There are 400 and some change, total CIDs. Part 5 was 500 rows
         // as opensecrets.org only allows 200 downloads per day.
         var httpStatus = HttpStatus.OK;
-        var sectorResponses = candidatesApiManager.getSectorsListResponse(part);
-        var response = candidatesApiManager.mapOpenSecretsResponseToSectors(sectorResponses).stream()
+        var sectorResponses = candidateApiManager.getSectorsListResponse(part);
+        var response = candidateApiManager.mapOpenSecretsResponseToSectors(sectorResponses).stream()
                 .map(SectorResponse::new)
                 .toList();
         return new ResponseEntity<>(response, httpStatus);
@@ -41,8 +41,8 @@ public class AdminOpensecretsController {
         // There are 400 and some change, total CIDs. Part 5 was 500 rows
         // as opensecrets.org only allows 200 downloads per day.
         var httpStatus = HttpStatus.OK;
-        var contributorResponses = candidatesApiManager.getContributorsListResponse(part);
-        var response = candidatesApiManager.mapOpenSecretsResponseToContributors(contributorResponses).stream()
+        var contributorResponses = candidateApiManager.getContributorsListResponse(part);
+        var response = candidateApiManager.mapOpenSecretsResponseToContributors(contributorResponses).stream()
                 .map(ContributorResponse::new)
                 .toList();
         return new ResponseEntity<>(response, httpStatus);
@@ -54,7 +54,7 @@ public class AdminOpensecretsController {
     public ResponseEntity<CampaignResponse> downloadCampaign(@RequestBody CommitteeRequest data) {
         var response = new CampaignResponse();
         var httpStatus = HttpStatus.OK;
-        response = candidatesApiManager.getPresidentialCampaignListResponse(data);
+        response = candidateApiManager.getPresidentialCampaignListResponse(data);
 
         return new ResponseEntity<>(response, httpStatus);
     }
