+++ src/main/java/co/inajar/oursponsors/dbos/repos/PreferenceRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PreferencesRepo extends JpaRepository<Preferences, Long> {
+public interface PreferenceRepo extends JpaRepository<Preferences, Long> {
     Preferences findPreferencesByUserId(Long id);
 
 }
