+++ src/main/java/co/inajar/oursponsors/models/propublica/congress/MiniCongressResponse.java
@@ -3,7 +3,7 @@ package co.inajar.oursponsors.models.propublica.congress;
 import co.inajar.oursponsors.dbos.entities.chambers.Congress;
 import co.inajar.oursponsors.models.fec.SponsorRequest;
 import co.inajar.oursponsors.models.fec.SponsorResponse;
-import co.inajar.oursponsors.services.fec.SponsorsManager;
+import co.inajar.oursponsors.services.fec.SponsorManager;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
@@ -20,7 +20,7 @@ public class MiniCongressResponse {
     private static final String BASE_URL = "https://theunitedstates.io/images/congress/original";
     @JsonProperty(value = "rep_type")
     private String repType;
-    private SponsorsManager sponsorsManager;
+    private SponsorManager sponsorManager;
     @JsonProperty(value = "id")
     private Long id;
     @JsonProperty(value = "title")
@@ -38,7 +38,7 @@ public class MiniCongressResponse {
     @JsonProperty(value = "sponsors")
     private Set<SponsorResponse> sponsors;
 
-    public MiniCongressResponse(Congress congress, SponsorsManager sponsorsManager) {
+    public MiniCongressResponse(Congress congress, SponsorManager sponsorManager) {
         repType = REPTYPE;
         id = congress.getId();
         title = congress.getTitle();
@@ -47,7 +47,7 @@ public class MiniCongressResponse {
         party = congress.getParty();
         state = congress.getState();
         imageUrl = BASE_URL + "/" + congress.getProPublicaId() + ".jpg";
-        this.sponsorsManager = sponsorsManager;
+        this.sponsorManager = sponsorManager;
 
         sponsors = new HashSet<>();
 
@@ -55,7 +55,7 @@ public class MiniCongressResponse {
         data.setChamber("congress");
         data.setOsId(congress.getId());
 
-        var possibleSponsors = Optional.ofNullable(sponsorsManager.getSponsors(data));
+        var possibleSponsors = Optional.ofNullable(sponsorManager.getSponsors(data));
         if (possibleSponsors.isPresent()) {
             var set = possibleSponsors.get().parallelStream()
                     .map(SponsorResponse::new)
