+++ src/main/java/co/inajar/oursponsors/services/propublica/MemberManager.java
@@ -10,7 +10,7 @@ import co.inajar.oursponsors.models.propublica.senator.SenatorDetailsResponse;
 import java.util.List;
 import java.util.Optional;
 
-public interface MembersManager {
+public interface MemberManager {
 
     Optional<List<Senator>> getSenators(User user);
